#93 - split watchers in aspects and outgoing components. integrated the RestTemplate handling into the existing RequestScope handler to provide exception and latency assaults.
