+++ chaos-monkey-spring-boot/src/main/java/de/codecentric/spring/boot/chaos/monkey/configuration/ChaosMonkeyConfiguration.java
@@ -35,16 +35,14 @@ import de.codecentric.spring.boot.chaos.monkey.configuration.toggles.DefaultChao
 import de.codecentric.spring.boot.chaos.monkey.configuration.toggles.DefaultChaosToggles;
 import de.codecentric.spring.boot.chaos.monkey.endpoints.ChaosMonkeyJmxEndpoint;
 import de.codecentric.spring.boot.chaos.monkey.endpoints.ChaosMonkeyRestEndpoint;
-import de.codecentric.spring.boot.chaos.monkey.watcher.SpringComponentAspect;
-import de.codecentric.spring.boot.chaos.monkey.watcher.SpringControllerAspect;
-import de.codecentric.spring.boot.chaos.monkey.watcher.SpringRepositoryAspectJDBC;
-import de.codecentric.spring.boot.chaos.monkey.watcher.SpringRepositoryAspectJPA;
-import de.codecentric.spring.boot.chaos.monkey.watcher.SpringRestControllerAspect;
-import de.codecentric.spring.boot.chaos.monkey.watcher.SpringServiceAspect;
-import de.codecentric.spring.boot.chaos.monkey.web.client.ChaosMonkeyRestTemplateInterceptor;
-import de.codecentric.spring.boot.chaos.monkey.web.client.ChaosMonkeyRestTemplatePostProcessor;
-import de.codecentric.spring.boot.chaos.monkey.web.client.RestTemplateAssault;
-import de.codecentric.spring.boot.chaos.monkey.web.client.RestTemplateErrorResponseAssault;
+import de.codecentric.spring.boot.chaos.monkey.watcher.aspect.SpringComponentAspect;
+import de.codecentric.spring.boot.chaos.monkey.watcher.aspect.SpringControllerAspect;
+import de.codecentric.spring.boot.chaos.monkey.watcher.aspect.SpringRepositoryAspectJDBC;
+import de.codecentric.spring.boot.chaos.monkey.watcher.aspect.SpringRepositoryAspectJPA;
+import de.codecentric.spring.boot.chaos.monkey.watcher.aspect.SpringRestControllerAspect;
+import de.codecentric.spring.boot.chaos.monkey.watcher.aspect.SpringServiceAspect;
+import de.codecentric.spring.boot.chaos.monkey.watcher.outgoing.ChaosMonkeyRestTemplatePostProcessor;
+import de.codecentric.spring.boot.chaos.monkey.watcher.outgoing.ChaosMonkeyRestTemplateWatcher;
 import java.io.IOException;
 import java.nio.charset.Charset;
 import java.util.List;
@@ -72,9 +70,9 @@ import org.springframework.util.StreamUtils;
 @Configuration
 @Profile("chaos-monkey")
 @EnableConfigurationProperties({
-    ChaosMonkeyProperties.class,
-    AssaultProperties.class,
-    WatcherProperties.class
+  ChaosMonkeyProperties.class,
+  AssaultProperties.class,
+  WatcherProperties.class
 })
 @Import(UnleashChaosConfiguration.class)
 @EnableScheduling
@@ -249,44 +247,29 @@ public class ChaosMonkeyConfiguration {
   }
 
   @Configuration
-  @ConditionalOnProperty(
-      prefix = "chaos.monkey.web.client",
-      value = "enabled",
-      havingValue = "true")
-  @EnableConfigurationProperties(
-      WebClientAssaultProperties.class
-  )
   static class ChaosMonkeyWebClientConfiguration {
 
     @Configuration
     @ConditionalOnProperty(
-        prefix = "chaos.monkey.web.client.rest-template",
-        value = "enabled",
+        prefix = "chaos.monkey.watcher",
+        value = "rest-template",
         havingValue = "true")
     static class ChaosMonkeyRestTemplateConfiguration {
 
       @Bean
       public ChaosMonkeyRestTemplatePostProcessor chaosMonkeyRestTemplatePostProcessor(
-          final ChaosMonkeyRestTemplateInterceptor chaosMonkeyRestTemplateInterceptor) {
-        return new ChaosMonkeyRestTemplatePostProcessor(chaosMonkeyRestTemplateInterceptor);
+          final ChaosMonkeyRestTemplateWatcher chaosMonkeyRestTemplateWatcher) {
+        return new ChaosMonkeyRestTemplatePostProcessor(chaosMonkeyRestTemplateWatcher);
       }
 
       @Bean
-      public ChaosMonkeyRestTemplateInterceptor chaosMonkeyRestTemplateInterceptor(
-          final List<RestTemplateAssault> restTemplateAssaults,
-          final WebClientAssaultProperties properties) {
-        return new ChaosMonkeyRestTemplateInterceptor(restTemplateAssaults, properties);
-      }
-
-      @Bean
-      @ConditionalOnProperty(
-          prefix = "chaos.monkey.web.client.rest-template.",
-          value = "error-response",
-          havingValue = "true")
-      public RestTemplateAssault restTemplateErrorResponseAssault() {
-        return new RestTemplateErrorResponseAssault();
+      public ChaosMonkeyRestTemplateWatcher chaosMonkeyRestTemplateInterceptor(
+          final ChaosMonkeyRequestScope chaosMonkeyRequestScope,
+          final WatcherProperties watcherProperties,
+          final AssaultProperties assaultProperties) {
+        return new ChaosMonkeyRestTemplateWatcher(
+            chaosMonkeyRequestScope, watcherProperties, assaultProperties);
       }
     }
-
   }
 }
