+++ chaos-monkey-spring-boot/src/main/java/de/codecentric/spring/boot/chaos/monkey/watcher/aspect/SpringControllerAspect.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package de.codecentric.spring.boot.chaos.monkey.watcher;
+package de.codecentric.spring.boot.chaos.monkey.watcher.aspect;
 
 import de.codecentric.spring.boot.chaos.monkey.component.ChaosMonkeyRequestScope;
 import de.codecentric.spring.boot.chaos.monkey.component.ChaosTarget;
@@ -33,7 +33,7 @@ import org.aspectj.lang.reflect.MethodSignature;
 @Aspect
 @AllArgsConstructor
 @Slf4j
-public class SpringRestControllerAspect extends ChaosMonkeyBaseAspect {
+public class SpringControllerAspect extends ChaosMonkeyBaseAspect {
 
   private final ChaosMonkeyRequestScope chaosMonkeyRequestScope;
 
@@ -41,25 +41,24 @@ public class SpringRestControllerAspect extends ChaosMonkeyBaseAspect {
 
   private WatcherProperties watcherProperties;
 
-  @Pointcut("within(@org.springframework.web.bind.annotation.RestController *)")
+  @Pointcut("within(@org.springframework.stereotype.Controller *)")
   public void classAnnotatedWithControllerPointcut() {}
 
   @Around(
       "classAnnotatedWithControllerPointcut() && allPublicMethodPointcut() && !classInChaosMonkeyPackage()")
   public Object intercept(ProceedingJoinPoint pjp) throws Throwable {
 
-    if (watcherProperties.isRestController()) {
-      log.debug("Watching public method on rest controller class: {}", pjp.getSignature());
+    if (watcherProperties.isController()) {
+      log.debug("Watching public method on controller class: {}", pjp.getSignature());
 
       if (metricEventPublisher != null) {
         metricEventPublisher.publishMetricEvent(
-            calculatePointcut(pjp.toShortString()), MetricType.RESTCONTROLLER);
+            calculatePointcut(pjp.toShortString()), MetricType.CONTROLLER);
       }
 
       MethodSignature signature = (MethodSignature) pjp.getSignature();
 
-      chaosMonkeyRequestScope.callChaosMonkey(
-          ChaosTarget.REST_CONTROLLER, createSignature(signature));
+      chaosMonkeyRequestScope.callChaosMonkey(ChaosTarget.CONTROLLER, createSignature(signature));
     }
     return pjp.proceed();
   }
