--- chaos-monkey-spring-boot/src/main/java/de/codecentric/spring/boot/chaos/monkey/web/client/RestTemplateAssault.java
@@ -1,12 +1,0 @@
-package de.codecentric.spring.boot.chaos.monkey.web.client;
-
-import org.springframework.http.HttpRequest;
-import org.springframework.http.client.ClientHttpRequestExecution;
-import org.springframework.http.client.ClientHttpResponse;
-
-public interface RestTemplateAssault {
-
-  ClientHttpResponse process(HttpRequest httpRequest,
-      ClientHttpRequestExecution clientHttpRequestExecution);
-
-}
