+++ chaos-monkey-spring-boot/src/test/java/de/codecentric/spring/boot/chaos/monkey/watcher/aspect/SpringRepositoryAspectJDBCTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 the original author or authors.
+ * Copyright 2019 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package de.codecentric.spring.boot.chaos.monkey.watcher;
+package de.codecentric.spring.boot.chaos.monkey.watcher.aspect;
 
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -25,18 +25,18 @@ import de.codecentric.spring.boot.chaos.monkey.component.ChaosTarget;
 import de.codecentric.spring.boot.chaos.monkey.component.MetricEventPublisher;
 import de.codecentric.spring.boot.chaos.monkey.component.MetricType;
 import de.codecentric.spring.boot.chaos.monkey.configuration.WatcherProperties;
-import de.codecentric.spring.boot.demo.chaos.monkey.restcontroller.DemoRestController;
+import de.codecentric.spring.boot.demo.chaos.monkey.repository.DemoRepositoryJDBC;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.aop.aspectj.annotation.AspectJProxyFactory;
 
-/** @author Benjamin Wilms */
+/** @author Eric Wyles */
 @ExtendWith(MockitoExtension.class)
-class SpringRestControllerAspectTest {
+class SpringRepositoryAspectJDBCTest {
 
-  private DemoRestController target = new DemoRestController();
+  private DemoRepositoryJDBC target = new DemoRepositoryJDBC();
 
   private WatcherProperties watcherProperties = new WatcherProperties();
 
@@ -46,14 +46,14 @@ class SpringRestControllerAspectTest {
 
   @Mock private MetricEventPublisher metricsMock;
 
-  private String pointcutName = "execution.DemoRestController.sayHello";
+  private String pointcutName = "execution.DemoRepositoryJDBC.sayHello";
 
   private String simpleName =
-      "de.codecentric.spring.boot.demo.chaos.monkey.restcontroller.DemoRestController.sayHello";
+      "de.codecentric.spring.boot.demo.chaos.monkey.repository.DemoRepositoryJDBC.sayHello";
 
   @Test
   void chaosMonkeyIsCalledWhenEnabledInConfig() {
-    watcherProperties.setRestController(true);
+    watcherProperties.setRepository(true);
 
     addRelevantAspect();
 
@@ -64,7 +64,7 @@ class SpringRestControllerAspectTest {
 
   @Test
   void chaosMonkeyIsNotCalledWhenDisabledInConfig() {
-    watcherProperties.setController(false);
+    watcherProperties.setRepository(false);
 
     addRelevantAspect();
 
@@ -89,39 +89,39 @@ class SpringRestControllerAspectTest {
   }
 
   private void addRelevantAspect() {
-    SpringRestControllerAspect restControllerAspect =
-        new SpringRestControllerAspect(chaosMonkeyRequestScopeMock, metricsMock, watcherProperties);
-    factory.addAspect(restControllerAspect);
+    SpringRepositoryAspectJDBC repositoryAspectJDBC =
+        new SpringRepositoryAspectJDBC(chaosMonkeyRequestScopeMock, metricsMock, watcherProperties);
+    factory.addAspect(repositoryAspectJDBC);
   }
 
   private void addNonRelevantAspects() {
+    SpringServiceAspect serviceAspect =
+        new SpringServiceAspect(chaosMonkeyRequestScopeMock, metricsMock, watcherProperties);
     SpringControllerAspect controllerAspect =
         new SpringControllerAspect(chaosMonkeyRequestScopeMock, metricsMock, watcherProperties);
     SpringComponentAspect componentAspect =
         new SpringComponentAspect(chaosMonkeyRequestScopeMock, metricsMock, watcherProperties);
-    SpringServiceAspect serviceAspect =
-        new SpringServiceAspect(chaosMonkeyRequestScopeMock, metricsMock, watcherProperties);
+    SpringRestControllerAspect restControllerAspect =
+        new SpringRestControllerAspect(chaosMonkeyRequestScopeMock, metricsMock, watcherProperties);
     SpringRepositoryAspectJPA repositoryAspect =
         new SpringRepositoryAspectJPA(chaosMonkeyRequestScopeMock, metricsMock, watcherProperties);
-    SpringRepositoryAspectJDBC repositoryStereotypeAspect =
-        new SpringRepositoryAspectJDBC(chaosMonkeyRequestScopeMock, metricsMock, watcherProperties);
 
+    factory.addAspect(serviceAspect);
     factory.addAspect(controllerAspect);
     factory.addAspect(componentAspect);
-    factory.addAspect(serviceAspect);
+    factory.addAspect(restControllerAspect);
     factory.addAspect(repositoryAspect);
-    factory.addAspect(repositoryStereotypeAspect);
   }
 
   private void callTargetMethod() {
-    DemoRestController proxy = factory.getProxy();
+    DemoRepositoryJDBC proxy = factory.getProxy();
     proxy.sayHello();
   }
 
   private void verifyDependenciesCalledXTimes(int i) {
     verify(chaosMonkeyRequestScopeMock, times(i))
-        .callChaosMonkey(ChaosTarget.REST_CONTROLLER, simpleName);
-    verify(metricsMock, times(i)).publishMetricEvent(pointcutName, MetricType.RESTCONTROLLER);
+        .callChaosMonkey(ChaosTarget.REPOSITORY, simpleName);
+    verify(metricsMock, times(i)).publishMetricEvent(pointcutName, MetricType.REPOSITORY);
     verifyNoMoreInteractions(chaosMonkeyRequestScopeMock, metricsMock);
   }
 }
