+++ chaos-monkey-spring-boot/src/test/java/de/codecentric/spring/boot/chaos/monkey/watcher/aspect/SpringServiceAspectTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package de.codecentric.spring.boot.chaos.monkey.watcher;
+package de.codecentric.spring.boot.chaos.monkey.watcher.aspect;
 
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -25,7 +25,7 @@ import de.codecentric.spring.boot.chaos.monkey.component.ChaosTarget;
 import de.codecentric.spring.boot.chaos.monkey.component.MetricEventPublisher;
 import de.codecentric.spring.boot.chaos.monkey.component.MetricType;
 import de.codecentric.spring.boot.chaos.monkey.configuration.WatcherProperties;
-import de.codecentric.spring.boot.demo.chaos.monkey.controller.DemoController;
+import de.codecentric.spring.boot.demo.chaos.monkey.service.DemoService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
@@ -34,9 +34,9 @@ import org.springframework.aop.aspectj.annotation.AspectJProxyFactory;
 
 /** @author Benjamin Wilms */
 @ExtendWith(MockitoExtension.class)
-class SpringControllerAspectTest {
+class SpringServiceAspectTest {
 
-  private DemoController target = new DemoController();
+  private DemoService target = new DemoService();
 
   private WatcherProperties watcherProperties = new WatcherProperties();
 
@@ -46,14 +46,14 @@ class SpringControllerAspectTest {
 
   @Mock private MetricEventPublisher metricsMock;
 
-  private String pointcutName = "execution.DemoController.sayHello";
+  private String pointcutName = "execution.DemoService.sayHello";
 
   private String simpleName =
-      "de.codecentric.spring.boot.demo.chaos.monkey.controller.DemoController.sayHello";
+      "de.codecentric.spring.boot.demo.chaos.monkey.service.DemoService.sayHello";
 
   @Test
   void chaosMonkeyIsCalledWhenEnabledInConfig() {
-    watcherProperties.setController(true);
+    watcherProperties.setService(true);
 
     addRelevantAspect();
 
@@ -64,7 +64,7 @@ class SpringControllerAspectTest {
 
   @Test
   void chaosMonkeyIsNotCalledWhenDisabledInConfig() {
-    watcherProperties.setController(false);
+    watcherProperties.setService(false);
 
     addRelevantAspect();
 
@@ -89,14 +89,14 @@ class SpringControllerAspectTest {
   }
 
   private void addRelevantAspect() {
-    SpringControllerAspect controllerAspect =
-        new SpringControllerAspect(chaosMonkeyRequestScopeMock, metricsMock, watcherProperties);
-    factory.addAspect(controllerAspect);
+    SpringServiceAspect serviceAspect =
+        new SpringServiceAspect(chaosMonkeyRequestScopeMock, metricsMock, watcherProperties);
+    factory.addAspect(serviceAspect);
   }
 
   private void addNonRelevantAspects() {
-    SpringServiceAspect serviceAspect =
-        new SpringServiceAspect(chaosMonkeyRequestScopeMock, metricsMock, watcherProperties);
+    SpringControllerAspect controllerAspect =
+        new SpringControllerAspect(chaosMonkeyRequestScopeMock, metricsMock, watcherProperties);
     SpringComponentAspect componentAspect =
         new SpringComponentAspect(chaosMonkeyRequestScopeMock, metricsMock, watcherProperties);
     SpringRestControllerAspect restControllerAspect =
@@ -106,7 +106,7 @@ class SpringControllerAspectTest {
     SpringRepositoryAspectJDBC repositoryStereotypeAspect =
         new SpringRepositoryAspectJDBC(chaosMonkeyRequestScopeMock, metricsMock, watcherProperties);
 
-    factory.addAspect(serviceAspect);
+    factory.addAspect(controllerAspect);
     factory.addAspect(componentAspect);
     factory.addAspect(restControllerAspect);
     factory.addAspect(repositoryAspect);
@@ -114,14 +114,13 @@ class SpringControllerAspectTest {
   }
 
   private void callTargetMethod() {
-    DemoController proxy = factory.getProxy();
+    DemoService proxy = factory.getProxy();
     proxy.sayHello();
   }
 
   private void verifyDependenciesCalledXTimes(int i) {
-    verify(chaosMonkeyRequestScopeMock, times(i))
-        .callChaosMonkey(ChaosTarget.CONTROLLER, simpleName);
-    verify(metricsMock, times(i)).publishMetricEvent(pointcutName, MetricType.CONTROLLER);
+    verify(chaosMonkeyRequestScopeMock, times(i)).callChaosMonkey(ChaosTarget.SERVICE, simpleName);
+    verify(metricsMock, times(i)).publishMetricEvent(pointcutName, MetricType.SERVICE);
     verifyNoMoreInteractions(chaosMonkeyRequestScopeMock, metricsMock);
   }
 }
