+++ chaos-monkey-spring-boot/src/test/java/de/codecentric/spring/boot/demo/chaos/monkey/ChaosDemoApplication.java
@@ -16,15 +16,21 @@
 
 package de.codecentric.spring.boot.demo.chaos.monkey;
 
+import de.codecentric.spring.boot.demo.chaos.monkey.ChaosDemoApplication.TestRestTemplateConfigurationProperties;
+import java.time.Duration;
+import java.time.temporal.ChronoUnit;
+import lombok.Data;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.web.client.RestTemplate;
 
-/**
- * @author Benjamin Wilms
- */
+/** @author Benjamin Wilms */
 @SpringBootApplication
+@EnableConfigurationProperties(value = {TestRestTemplateConfigurationProperties.class})
 public class ChaosDemoApplication {
 
   public static void main(String[] args) {
@@ -32,7 +38,15 @@ public class ChaosDemoApplication {
   }
 
   @Bean
-  public RestTemplate restTemplate() {
-    return new RestTemplate();
+  public RestTemplate restTemplate(final TestRestTemplateConfigurationProperties properties) {
+    return new RestTemplateBuilder()
+        .setReadTimeout(Duration.of(properties.timeOut, ChronoUnit.MILLIS))
+        .build();
+  }
+
+  @Data
+  @ConfigurationProperties(prefix = "chaos.monkey.test.rest-template")
+  static class TestRestTemplateConfigurationProperties {
+    private Long timeOut = 10000L;
   }
 }
