+++ chaos-monkey-spring-boot/src/test/java/de/codecentric/spring/boot/demo/chaos/monkey/service/DemoRestTemplateService.java
@@ -14,7 +14,6 @@ public class DemoRestTemplateService {
   }
 
   public ResponseEntity<?> callWithRestTemplate() {
-    return this.restTemplate
-        .getForEntity("https://www.google.de", String.class);
+    return this.restTemplate.getForEntity("https://www.google.de", String.class);
   }
 }
